pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cybersecurity-threat-detection"
version = "1.0.0"
description = "OpenEnv cybersecurity incident response environment"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Security AI Team" }
]
license = { text = "MIT" }
dependencies = [
  "openai>=1.3.0",
  "pydantic>=2.0.0",
  "PyYAML>=6.0",
  "python-dotenv>=1.0.0",
  "Flask>=3.0.0",
  "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "tasks*", "openenv*"]
